+++ src/main/java/de/codescape/jira/plugins/scrumpoker/condition/ScrumPokerForIssueCondition.java
@@ -6,17 +6,24 @@ import com.atlassian.jira.issue.fields.CustomField;
 import com.atlassian.jira.plugin.webfragment.conditions.AbstractIssueWebCondition;
 import com.atlassian.jira.plugin.webfragment.model.JiraHelper;
 import com.atlassian.jira.user.ApplicationUser;
+import com.atlassian.plugin.spring.scanner.annotation.component.Scanned;
 import de.codescape.jira.plugins.scrumpoker.service.EstimationFieldService;
+
+import javax.inject.Inject;
+import javax.inject.Named;
 
 /**
  * This condition is used to decide whether a button to start a Scrum poker session should be displayed or not for the
  * given issue. The issue must be editable and have the custom field that is configured for Scrum poker estimations.
  */
+@Named
+@Scanned
 public class ScrumPokerForIssueCondition extends AbstractIssueWebCondition {
 
     private final CustomFieldManager customFieldManager;
     private final EstimationFieldService estimationFieldService;
 
+    @Inject
     public ScrumPokerForIssueCondition(CustomFieldManager customFieldManager,
                                        EstimationFieldService estimationFieldService) {
         this.customFieldManager = customFieldManager;
@@ -25,6 +32,13 @@ public class ScrumPokerForIssueCondition extends AbstractIssueWebCondition {
 
     @Override
     public boolean shouldDisplay(ApplicationUser applicationUser, Issue issue, JiraHelper jiraHelper) {
+        return isEstimable(issue);
+    }
+
+    /**
+     * Returns whether a Scrum poker session can be started for the given issue.
+     */
+    public boolean isEstimable(Issue issue) {
         return issue.isEditable() && hasStoryPointField(issue);
     }