+++ src/test/java/de/codescape/jira/plugins/scrumpoker/condition/ScrumPokerForIssueConditionTest.java
@@ -6,7 +6,6 @@ import com.atlassian.jira.issue.fields.CustomField;
 import com.atlassian.jira.plugin.webfragment.model.JiraHelper;
 import com.atlassian.jira.user.ApplicationUser;
 import de.codescape.jira.plugins.scrumpoker.service.EstimationFieldService;
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.mockito.InjectMocks;
@@ -15,7 +14,6 @@ import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 
 import static org.hamcrest.Matchers.is;
@@ -51,36 +49,24 @@ public class ScrumPokerForIssueConditionTest {
     @Mock
     private Issue issue;
 
-    @Before
-    public void before() {
-        // configured field for Scrum Poker
-        when(estimationFieldService.findStoryPointField()).thenReturn(storyPointField);
-
-        // expected ids for fields
-        when(storyPointField.getId()).thenReturn("STORY_POINT_FIELD_ID");
-        when(someOtherField.getId()).thenReturn("SOME_OTHER_FIELD_ID");
-
-        when(jiraHelper.getContextParams()).thenReturn(mapThatContainsTheIssue());
-    }
-
     @Test
     public void shouldDisplayForEditableIssueWithStoryPointField() {
         expectThatIssueContainsTheStoryPointField();
         expectThatIssueIsEditable();
-        assertThat(scrumPokerForIssueCondition.shouldDisplay(applicationUser, jiraHelper), is(true));
+        assertThat(scrumPokerForIssueCondition.shouldDisplay(applicationUser, issue, jiraHelper), is(true));
+    }
+
+    @Test
+    public void shouldNotDisplayForNonEditableIssue() {
+        expectThatIssueIsNotEditable();
+        assertThat(scrumPokerForIssueCondition.shouldDisplay(applicationUser, issue, jiraHelper), is(false));
     }
 
     @Test
     public void shouldNotDisplayForEditableIssueWithoutStoryPointField() {
         expectThatIssueDoesNotContainTheStoryPointField();
         expectThatIssueIsEditable();
-        assertThat(scrumPokerForIssueCondition.shouldDisplay(applicationUser, jiraHelper), is(false));
-    }
-
-    @Test
-    public void shouldNotDisplayForNonEditableIssue() {
-        expectThatIssueIsNotEditable();
-        assertThat(scrumPokerForIssueCondition.shouldDisplay(applicationUser, jiraHelper), is(false));
+        assertThat(scrumPokerForIssueCondition.shouldDisplay(applicationUser, issue, jiraHelper), is(false));
     }
 
     private void expectThatIssueIsNotEditable() {
@@ -92,22 +78,22 @@ public class ScrumPokerForIssueConditionTest {
     }
 
     private void expectThatIssueDoesNotContainTheStoryPointField() {
+        expectThatEstimationFieldExists();
         List<CustomField> emptyList = new ArrayList<>();
         emptyList.add(someOtherField);
         when(customFieldManager.getCustomFieldObjects(issue)).thenReturn(emptyList);
     }
 
     private void expectThatIssueContainsTheStoryPointField() {
+        expectThatEstimationFieldExists();
         ArrayList<CustomField> listOfCustomFields = new ArrayList<>();
         listOfCustomFields.add(storyPointField);
         listOfCustomFields.add(someOtherField);
         when(customFieldManager.getCustomFieldObjects(issue)).thenReturn(listOfCustomFields);
     }
 
-    private HashMap<String, Object> mapThatContainsTheIssue() {
-        HashMap<String, Object> contextParams = new HashMap<>();
-        contextParams.put("issue", issue);
-        return contextParams;
+    private void expectThatEstimationFieldExists() {
+        when(estimationFieldService.findStoryPointField()).thenReturn(storyPointField);
     }
 
 }