+++ changelog.md
@@ -9,6 +9,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 * refactor: rename Scrum poker session service implementation
 * feat: prevent issue description from floating into poker elements
 * refactor: change primary key for persisted votes from int to Long
+* feat: do not create Scrum poker sessions for issues missing estimation field
 
 ## [3.3.0] - 2018-09-01